requirements.txt
tensorflow==2.19.0
tensorflow-datasets
streamlit
numpy
pillow
